+++ src/cli/html-validate.ts
@@ -1,8 +1,8 @@
 import HtmlValidate from '../htmlvalidate';
-import { Result } from '../reporter'; // eslint-disable-line no-unused-vars
+import { Result, Reporter } from '../reporter';
+import * as minimist from 'minimist';
 
 const fs = require('fs');
-const minimist = require('minimist');
 const glob = require('glob');
 
 function getMode(argv: { [key: string]: any }){
@@ -55,7 +55,7 @@ function getFormatters(formatters: string): ((results: Result[]) => void)[] {
 	});
 }
 
-const argv = minimist(process.argv.slice(2), {
+const argv: minimist.ParsedArgs = minimist(process.argv.slice(2), {
 	string: ['f', 'formatter', 'rule'],
 	boolean: ['dump-events', 'dump-tokens', 'dump-tree'],
 	alias: {
@@ -101,18 +101,12 @@ const config = getGlobalConfig(argv.rule);
 const formatters = getFormatters(argv.formatter);
 const htmlvalidate = new HtmlValidate(config);
 
-let results: Result[] = [];
-let valid = true;
+const files = argv._.reduce((files: string[], pattern: string) => {
+	return files.concat(glob.sync(pattern));
+}, []);
+const unique = [... new Set(files)];
+const results = unique.map((filename: string) => htmlvalidate.file(filename, mode));
+const merged = Reporter.merge(results);
 
-argv._.forEach((pattern: string) => {
-	glob.sync(pattern).forEach((filename: string) => {
-		const report = htmlvalidate.file(filename, mode);
-
-		/* aggregate results */
-		valid = valid && report.valid;
-		results = results.concat(report.results);
-	});
-});
-
-formatters.forEach((formatter: any) => formatter(results));
-process.exit(valid ? 0 : 1);
+formatters.forEach((formatter: any) => formatter(merged.results));
+process.exit(merged.valid ? 0 : 1);